moved default preprocessor to connectome group
